backend/requirement.txt
Django==3.2.4
pandas==1.3.3
matplotlib==3.4.3
djangorestframework==3.14.0
requests==2.31.0